neighborhood-intelligence/requirements.txt
# Core dependencies
streamlit>=1.28.0
google-generativeai>=0.3.0
googlemaps>=4.10.0
firecrawl-py>=0.0.16
duckduckgo-search>=4.1.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# Optional: for backup scraping
playwright>=1.40.0

# Development
pytest>=7.4.0
